requirements.txt
hidapi>=0.14.0
psutil>=5.9.0
PyYAML>=6.0
Flask>=2.3.0
